@@ -8,30 +8,32 @@ class User < ActiveRecord::Base
 
   has_and_belongs_to_many :roles
 
-  has_many :test_requests, :order => "submitted_at DESC"
+  #has_and_belongs_to_many :groups
+  has_many :groups_users, class_name: GroupUser
+  has_many :groups, :through => :groups_users
 
-  has_many :messages, 
-           :class_name => "Message",
-           :foreign_key => "sender_id", 
-           :order => 'created_at DESC'
+  has_many :test_requests, -> {order(submitted_at: DESC)}
 
-  has_many :replied_messages, 
+  has_many :messages, -> { order(created_at: DESC) },
            :class_name => "Message",
-           :foreign_key => "receiver_id", 
-           :order => 'created_at DESC'
+           :foreign_key => "sender_id"
+
+  has_many :replied_messages, -> { order(created_at: DESC) },
+           :class_name => "Message",
+           :foreign_key => "receiver_id"
 
   has_one :contest_stat, :class_name => "UserContestStat", :dependent => :destroy
 
   belongs_to :site
   belongs_to :country
 
-  has_and_belongs_to_many :contests, :uniq => true, :order => 'name'
+  has_and_belongs_to_many :contests, -> { order(:name); uniq}
 
-  scope :activated_users, :conditions => {:activated => true}
+  scope :activated_users, -> {where activated: true}
 
   validates_presence_of :login
   validates_uniqueness_of :login
-  validates_format_of :login, :with => /^[\_A-Za-z0-9]+$/
+  validates_format_of :login, :with => /\A[\_A-Za-z0-9]+\z/
   validates_length_of :login, :within => 3..30
 
   validates_presence_of :full_name
@@ -111,6 +113,7 @@ class User < ActiveRecord::Base
     begin
       http = Net::HTTP.new('www.cas.chula.ac.th', 443)
       http.use_ssl = true
+      http.verify_mode = OpenSSL::SSL::VERIFY_NONE
       result = [ ]
       http.start do |http|
         req = Net::HTTP::Post.new('/cas/api/?q=studentAuthenticate')
@@ -119,7 +122,7 @@ class User < ActiveRecord::Base
         result = JSON.parse resp.body
       end
       return true if result["type"] == "beanStudent"
-    rescue
+    rescue => e
       return false
     end
     return false
@@ -176,14 +179,14 @@ class User < ActiveRecord::Base
   end
 
   def self.find_non_admin_with_prefix(prefix='')
-    users = User.find(:all)
+    users = User.all
     return users.find_all { |u| !(u.admin?) and u.login.index(prefix)==0 }
   end
 
   # Contest information
 
   def self.find_users_with_no_contest()
-    users = User.find(:all)
+    users = User.all
     return users.find_all { |u| u.contests.length == 0 }
   end
 
@@ -239,7 +242,7 @@ class User < ActiveRecord::Base
 
   def update_start_time
     stat = self.contest_stat
-    if (stat.nil?) or (stat.started_at.nil?)
+    if stat.nil? or stat.started_at.nil?
       stat ||= UserContestStat.new(:user => self)
       stat.started_at = Time.now.gmtime
       stat.save
@@ -280,9 +283,23 @@ class User < ActiveRecord::Base
     return contest_problems
   end
 
+  def solve_all_available_problems?
+    available_problems.each do |p|
+      u = self
+      sub = Submission.find_last_by_user_and_problem(u.id,p.id)
+      return false if !p or !sub or sub.points < p.full_score
+    end
+    return true
+  end
+
+  #get a list of available problem
   def available_problems
     if not GraderConfiguration.multicontests?
-      return Problem.find_available_problems
+      if GraderConfiguration.use_problem_group?
+        return available_problems_in_group
+      else
+        return Problem.available_problems
+      end
     else
       contest_problems = []
       pin = {}
@@ -299,12 +316,32 @@ class User < ActiveRecord::Base
     end
   end
 
-  def can_view_problem?(problem)
-    if not GraderConfiguration.multicontests?
-      return problem.available
-    else
-      return problem_in_user_contests? problem
+  def available_problems_in_group
+    problem = []
+    self.groups.each do |group|
+      group.problems.where(available: true).each { |p| problem << p }
     end
+    problem.uniq!
+    if problem
+      problem.sort! do |a,b|
+        case
+        when a.date_added < b.date_added
+          1
+        when a.date_added > b.date_added
+          -1
+        else
+          a.name <=> b.name
+        end
+      end
+      return problem
+    else
+      return []
+    end
+  end
+
+  def can_view_problem?(problem)
+    return true if admin?
+    return available_problems.include? problem
   end
 
   def self.clear_last_login
