@@ -12,10 +12,14 @@
 //
 //= require jquery
 //= require jquery_ujs
-//= require jquery.ui.all
+//= require dataTables/jquery.dataTables
+//= require dataTables/bootstrap/3/jquery.dataTables.bootstrap
+//= require jquery-ui
 //= require bootstrap-sprockets
 //= require moment
+//= require moment/th
 //= require bootstrap-sortable
+//= require bootstrap-datetimepicker
 //= require select2
 //= require ace-rails-ap
 //= require ace/mode-c_cpp
@@ -28,13 +32,12 @@
 //= require custom
 //= require jquery.countdown
 //-------------- addition from local_jquery -----------
-//= require jquery.ui.datepicker
-//= require jquery.ui.slider
-//= require jquery-ui-timepicker-addon
 //= require jquery-tablesorter
 //= require best_in_place
 //= require best_in_place.jquery-ui
 //= require brython
+//= require bootstrap-datepicker
+//= require bootstrap-datetimepicker
 
 // since this is after blank line, it is not downloaded
 //x= require prototype
